feat: 新しいコンテキスト型を追加し、解析および対局設定の管理を強化
