+++ packages/ui/src/components/shogi-match/hooks/useGameControls.ts
@@ -293,7 +293,10 @@ export function useGameControls({
                 // NNUE未ダウンロードエラー → 評価関数ファイル管理を開いて理由を表示
                 const errorMessage =
                     e instanceof Error ? e.message : "評価関数の準備に失敗しました";
-                openNnueManager(`対局を開始できません: ${errorMessage}`);
+                const detailedMessage = `対局を開始できません: ${errorMessage}\n\n対処方法:\n1. NNUE管理画面が開きます\n2. エンジンに必要なファイルをダウンロードしてください\n3. 再度対局を開始してください`;
+                setMessage({ text: detailedMessage, type: "error" });
+                // NNUE管理ダイアログを開く
+                openNnueManager("missing-sente"); // または "missing-gote" を適切に判定
                 return;
             }
 
@@ -326,6 +329,7 @@ export function useGameControls({
         setIsEditMode,
         setPosition,
         openNnueManager,
+        setMessage,
     ]);
 
     /**
