+++ packages/ui/src/components/shogi-match/hooks/useMoveExecution.test-spec.md
@@ -1,0 +1,194 @@
+# useMoveExecution テスト要件
+
+このドキュメントは、`useMoveExecution`フックの包括的なテストケースを定義します。
+
+## テストカバレッジ目標
+
+以下の3つのモードと境界条件を網羅的にテストします：
+
+### 1. 編集モード
+
+#### 1.1 不正な手の許可
+- **テスト内容**: 編集モードでは合法手チェックをスキップし、任意の配置を許可する
+- **期待動作**: `fetchLegalMoves`が呼ばれず、`placePieceAt`が呼ばれる
+
+#### 1.2 駒の配置
+- **テスト内容**: 指定した駒を指定したマスに配置する
+- **期待動作**: `placePieceAt`が正しいパラメータで呼ばれる（owner, type, promoted）
+
+#### 1.3 駒の削除
+- **テスト内容**: 消しゴムツールで駒を削除する
+- **期待動作**: `placePieceAt(square, null)`が呼ばれる
+
+#### 1.4 駒の移動
+- **テスト内容**: 駒を別のマスに移動する
+- **期待動作**:
+  - 元のマスから駒を削除
+  - 新しいマスに駒を配置
+  - `setEditFromSquare(null)`で選択を解除
+
+#### 1.5 局面更新
+- **テスト内容**: 編集後の局面が正しく反映される
+- **期待動作**: `position`が更新される
+
+### 2. 対局モード
+
+#### 2.1 不正な手の拒否
+- **テスト内容**: 合法手リストに含まれない手を拒否する
+- **期待動作**: エラーメッセージが表示され、手が実行されない
+
+#### 2.2 合法手の実行
+- **テスト内容**: 合法手リストに含まれる手を実行する
+- **期待動作**:
+  - `navigation.addMove`が呼ばれる
+  - 局面が更新される
+  - 選択が解除される
+
+#### 2.3 エンジンの手番中の制限
+- **テスト内容**: エンジンが思考中は人間の手を受け付けない
+- **期待動作**: エラーメッセージが表示され、手が実行されない
+
+#### 2.4 時計の更新
+- **テスト内容**: 手を実行した後、時計が次の手番に更新される
+- **期待動作**: `updateClocksForNextTurn`が呼ばれる
+
+#### 2.5 自動対局のトリガー
+- **テスト内容**: 人間の手の後、エンジンの手番になった場合に自動対局が開始される
+- **期待動作**: エンジン管理が適切に処理される
+
+### 3. 検討モード
+
+#### 3.1 自由な手の探索
+- **テスト内容**: 検討モードでは任意の合法手を探索できる
+- **期待動作**:
+  - 合法手チェックが実行される
+  - 合法手であれば実行される
+
+#### 3.2 分岐の作成
+- **テスト内容**: 本譜と異なる手を指した場合、分岐が作成される
+- **期待動作**:
+  - `navigation.addBranch`が呼ばれる
+  - `setLastAddedBranchInfo`が呼ばれる
+
+#### 3.3 局面の復元
+- **テスト内容**: ナビゲーションで過去の局面に戻った場合、その局面が復元される
+- **期待動作**: `position`が正しく復元される
+
+### 4. 境界条件
+
+#### 4.1 成り処理
+
+##### 4.1.1 成り選択ダイアログの表示
+- **テスト内容**: 成りが可能な手（通常と成りの両方が合法手リストに含まれる）
+- **期待動作**: `setPromotionSelection`が呼ばれ、ダイアログが表示される
+
+##### 4.1.2 成り選択の確定
+- **テスト内容**: ダイアログで「成る」を選択
+- **期待動作**:
+  - 成りの手（末尾に`+`）が棋譜に追加される
+  - `setPromotionSelection(null)`でダイアログが閉じる
+
+##### 4.1.3 不成の選択
+- **テスト内容**: ダイアログで「成らない」を選択
+- **期待動作**:
+  - 通常の手が棋譜に追加される
+  - `setPromotionSelection(null)`でダイアログが閉じる
+
+##### 4.1.4 成りのキャンセル
+- **テスト内容**: ダイアログをキャンセル（ESCキーなど）
+- **期待動作**:
+  - 手は実行されない
+  - 選択が解除される
+
+#### 4.2 パス処理
+
+##### 4.2.1 パス手の実行
+- **テスト内容**: パス権がある状態でパスを実行
+- **期待動作**:
+  - パス手が棋譜に追加される
+  - パス権が減少する
+  - 手番が交代する
+
+##### 4.2.2 パス権なしの制限
+- **テスト内容**: パス権がない状態でパスを試みる
+- **期待動作**: エラーメッセージが表示され、パスできない
+
+#### 4.3 二重処理の防止
+
+##### 4.3.1 処理中フラグのチェック
+- **テスト内容**: 手の処理中に別の手を試みる
+- **期待動作**:
+  - `moveProcessingRef.current`が`true`の間は新しい手を受け付けない
+  - 警告メッセージが表示される（任意）
+
+##### 4.3.2 処理完了後のフラグリセット
+- **テスト内容**: 手の処理が完了した後
+- **期待動作**: `moveProcessingRef.current`が`false`に戻る
+
+#### 4.4 合法手キャッシュ
+
+##### 4.4.1 キャッシュヒット時
+- **テスト内容**: キャッシュに合法手が存在する場合
+- **期待動作**: `fetchLegalMoves`が呼ばれず、キャッシュから取得
+
+##### 4.4.2 キャッシュミス時
+- **テスト内容**: キャッシュに合法手が存在しない場合
+- **期待動作**: `fetchLegalMoves`が呼ばれ、結果がキャッシュに保存される
+
+##### 4.4.3 キャッシュのクリア
+- **テスト内容**: 局面が変更された場合
+- **期待動作**: `clearLegalCache`が呼ばれる
+
+#### 4.5 エラーハンドリング
+
+##### 4.5.1 合法手取得エラー
+- **テスト内容**: `fetchLegalMoves`がエラーをスローする
+- **期待動作**: エラーメッセージが表示され、手が実行されない
+
+##### 4.5.2 手の実行エラー
+- **テスト内容**: `navigation.addMove`がエラーをスローする
+- **期待動作**: エラーメッセージが表示され、状態がロールバックされる
+
+### 5. 持ち駒の処理
+
+#### 5.1 持ち駒の選択
+- **テスト内容**: 持ち駒を選択する
+- **期待動作**: `setSelection({ kind: "hand", piece })`が呼ばれる
+
+#### 5.2 持ち駒を打つ
+- **テスト内容**: 持ち駒を選択後、盤上のマスをクリック
+- **期待動作**:
+  - 打つ手が合法手リストに含まれる場合、実行される
+  - 持ち駒が減少する
+  - 手が棋譜に追加される
+
+#### 5.3 持ち駒の不正な打ち方
+- **テスト内容**: 既に駒が存在するマスに打つ
+- **期待動作**: エラーメッセージが表示され、手が実行されない
+
+## 実装時の注意事項
+
+1. **モックの準備**:
+   - `navigation`（棋譜ナビゲーション）
+   - `fetchLegalMoves`（合法手取得）
+   - `placePieceAt`（駒配置）
+   - 各種setter関数
+
+2. **非同期処理**:
+   - `handleSquareSelect`は非同期関数なので、`await`を使用
+   - `act`でラップする
+
+3. **状態の初期化**:
+   - 各テストケースで適切な初期状態を設定
+   - `beforeEach`でモックをクリア
+
+4. **テストフレームワーク**:
+   - Vitestを使用（`describe`, `it`, `expect`, `vi`）
+   - React Testing Libraryの`renderHook`を使用
+
+## 参考ファイル
+
+既存のテストファイルのパターンを参考にする：
+- `packages/ui/src/components/shogi-match/hooks/useClockManager.test.ts`
+- `packages/ui/src/components/shogi-match/hooks/useEngineManager.test.ts`
+- `packages/ui/src/components/shogi-match/hooks/usePositionEditor.test.ts`
